feat: add config to prevent dropping specified columns
Adds the config option `neverDropColumns` which specifies a list of
column names, and their associated model names, to prevent from being
dropped when generating database migrations.

The motivation for this is "hidden" properties. If you're using one
OpenAPI definition file to generate a backend API and a frontend SDK,
there may be some properties you want on the backend API (such as user
management/security columns, e.g. `auth_key`) that you don't want to
appear on the frontend SDK. With no clear way to do this in the OpenAPI
definition, the solution here is to simply not include the properties in
the definition and implement them manually in the backend API. When the
API code generator sees that the columns have been removed from the
definition, the new `neverDropColumns` config option is what prevents
them being dropped.