fix: migration name date conflict
When multiple migrations are generated separately (that is, by multiple
operations) in the same hour, the file names become unordered. This
causes migrations that depend on previous migrations to be generated out
of order, which has been causing migration failures.

This commit prevents this behaviour by adding the minutes and seconds to
the migration file name. The migration names are then ordered more
precisely.